debugging

+++ scripts/utils_sim.py
@@ -49,7 +49,7 @@ class arucoDetector():
         print('len(list_img) ', len(list_img))
         print('len(list_img 0 ) ', len(list_img[0]))
         for i in range(len(list_img)):
-            num_markers, id_markers = self.count_aruco_markers_from_single_image(list_img[i])
+            num_markers, id_markers = self.count_aruco_markers_from_single_image(list_img[i].astype(np.uint8))
             num_markers_in_img += num_markers
         return num_markers_in_img / len(list_img)
